extract point types
